// app/src/generated/eon-api/VedicAnalysisRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type VedicAnalysisRequest = { nowUtc: string | null, targetYear: number | null, year: number, month: number, day: number, hour: number, minute: number, isLunar: boolean, isLeapMonth: boolean, lat: number, lon: number, timezone: string, unknownTime: boolean | null, };