// app/src/generated/eon-api/TransitAnalysisRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type TransitAnalysisRequest = { nowUtc: string | null, isMale: boolean, useNightRatHour: boolean | null, year: number, month: number, day: number, hour: number, minute: number, isLunar: boolean, isLeapMonth: boolean, lat: number, lon: number, timezone: string, unknownTime: boolean | null, };